# Conflicts:
#	baystation12.dme
#	infinity/code/game/objects/items/weapons/storage/boxes.dm
#	infinity/icons/turf/flooring/decals.dmi

+++ code/modules/admin/view_variables/topic.dm
@@ -483,7 +483,8 @@
 
 		to_chat(usr, "Removed [rem_organ] from [M].")
 		rem_organ.removed()
-		qdel(rem_organ)
+		if(!QDELETED(rem_organ))
+			qdel(rem_organ)
 
 	else if(href_list["fix_nano"])
 		if(!check_rights(R_DEBUG)) return
@@ -530,7 +531,7 @@
 		switch(Text)
 			if("brute")	L.adjustBruteLoss(amount)
 			if("fire")	L.adjustFireLoss(amount)
-			if("toxin")	L.adjustToxLoss(amount)
+			if("toxin")	L.adjustToxLoss(amount, admin_healing = 1) //INF, WAS if("toxin")	L.adjustToxLoss(amount)
 			if("oxygen")L.adjustOxyLoss(amount)
 			if("brain")	L.adjustBrainLoss(amount)
 			if("clone")	L.adjustCloneLoss(amount)
